fix require var -> const and split test files
